Refactor :  책임 분리를 위한 HttpSevlet 인터페이스 추가
하나의 Servlet이 모든 요청을 처리하지 않도록 하기위해, 인터페이스를 추가하였습니다.

하지만 현재 코드에서는 과한 분리인 뜻한 느낌이 들었습니다.

그렇지만, 보기에는 좋은 것 같습니다.
