+++ src/main/java/servlet/HttpServlet.java
@@ -1,0 +1,8 @@
+package servlet;
+
+import model.HttpRequest;
+import model.HttpResponse;
+
+public interface HttpServlet {
+    public void service(HttpRequest httpRequest, HttpResponse httpResponse) throws Exception;
+}
